refactor: change structure of index.js file
