+++ portfolio/index.html
@@ -266,7 +266,7 @@
         </div>
     </footer>
 
-    <script src="index.js"></script>
+    <script type="module" src="index.js"></script>
 </body>
 
 </html>
